# Modify DecodeFrom to using scanner for Memory efficiency