--- a/AUTHORS
+++ b/AUTHORS
@@ -1,4 +1,4 @@
-There several persons contribute their code to the project.
+There are several persons contribute their code to the project.
 They listed below in an alphabetical order:
 
 - Alexander (Axel) I.Grafov <grafov@gmail.com>