--- a/AUTHORS
+++ b/AUTHORS
@@ -12,6 +12,7 @@ to the project. They listed below in an alphabetical order:
 - Kz26
 - Makombo
 - Scott Kidder <skidder@brightcove.com>
+- Vishal Kumar Tuniki <vishal24.tuniki@gmail.com>
 - Zac Shenker <zshenker@brightcove.com>
 
 If you want to be added to this list (or removed for any reason)