--- a/reader.go
+++ b/reader.go
@@ -518,8 +518,6 @@ func decodeLineOfMediaPlaylist(p *MediaPlaylist, wv *WV, state *decodingState, l
 		}
 		if len(params) > 1 {
 			title = params[1]
-		} else {
-			title = "<no title>"
 		}
 	case !state.tagDiscontinuity && strings.HasPrefix(line, "#EXT-X-DISCONTINUITY"):
 		state.tagDiscontinuity = true