- Modify DecodeFrom to using scanner for Memory efficiency.

Before:

```
BenchmarkDecodeMasterPlaylist-8    30000         46001 ns/op       12197 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-8       100      17027700 ns/op    11000965 B/op     120105 allocs/op
BenchmarkEncodeMasterPlaylist-8  1000000          1066 ns/op          40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-8       300       4794475 ns/op       45424 B/op        405 allocs/op
```

After:

```
BenchmarkDecodeMasterPlaylist-8    30000         45677 ns/op       14190 B/op         76 allocs/op
BenchmarkDecodeMediaPlaylist-8       100      16554687 ns/op     6812737 B/op     120089 allocs/op
BenchmarkEncodeMasterPlaylist-8  1000000          1081 ns/op          40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-8       300       4806393 ns/op       31466 B/op        405 allocs/op
```

Benchstat Summary

```
name                    old time/op    new time/op    delta
DecodeMasterPlaylist-8    46.0µs ± 0%    45.7µs ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     17.0ms ± 0%    16.6ms ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8    1.07µs ± 0%    1.08µs ± 0%   ~             (p=1.000 n=1+1)
EncodeMediaPlaylist-8     4.79ms ± 0%    4.81ms ± 0%   ~             (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
DecodeMasterPlaylist-8    12.2kB ± 0%    14.2kB ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     11.0MB ± 0%     6.8MB ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8     40.0B ± 0%     40.0B ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8     45.4kB ± 0%    31.5kB ± 0%   ~             (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
DecodeMasterPlaylist-8      77.0 ± 0%      76.0 ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8       120k ± 0%      120k ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8      11.0 ± 0%      11.0 ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8        405 ± 0%       405 ± 0%   ~     (all samples are equal)
```